rabbitmq/consumer/requirements.txt
pika==1.2.0
python-decouple==3.6
